# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberLab API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Username taken
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: New token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                required: [token]
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
...
  /findings:
    get:
      operationId: listFindings
      tags: [findings]
      summary: List findings
      security:
        - bearerAuth: []
      parameters:
        - name: targetId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Findings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
    post:
      operationId: createFinding
      tags: [findings]
      summary: Create a finding
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFindingRequest"
      responses:
        "201":
          description: Finding created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
  /findings/{id}:
    get:
      operationId: getFinding
      tags: [findings]
      summary: Get a finding
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
    patch:
      operationId: updateFinding
      tags: [findings]
      summary: Update finding status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFindingRequest"
      responses:
        "200":
          description: Updated finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
    delete:
      operationId: deleteFinding
      tags: [findings]
      summary: Delete a finding
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /findings/{id}/timeline:
    get:
      operationId: getFindingTimeline
      tags: [findings]
      summary: Get finding status timeline
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Finding events
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List lab notes
      security:
        - bearerAuth: []
      parameters:
        - name: targetId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a lab note
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required: [username, email, password]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, createdAt]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, refreshToken, user]
...
    Finding:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        targetId:
          type: string
        scanJobId:
          type: string
          nullable: true
        title:
          type: string
        cveId:
          type: string
          nullable: true
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        status:
          type: string
          enum: [open, fixed, accepted_risk, false_positive]
        cvssScore:
          type: number
          nullable: true
        owaspCategory:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        remediation:
          type: string
          nullable: true
        references:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, targetId, title, severity, status, references, createdAt, updatedAt]
    CreateFindingRequest:
      type: object
      properties:
        targetId:
          type: string
        scanJobId:
          type: string
        title:
          type: string
        cveId:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        cvssScore:
          type: number
        owaspCategory:
          type: string
        description:
          type: string
        remediation:
          type: string
        references:
          type: array
          items:
            type: string
        notes:
          type: string
      required: [targetId, title]
    UpdateFindingRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, fixed, accepted_risk, false_positive]
        notes:
          type: string
        remediation:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        note:
          type: string
    Note:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        targetId:
          type: string
        title:
          type: string
          nullable: true
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        attachmentUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, targetId, body, tags, attachmentUrls, createdAt, updatedAt]
    CreateNoteRequest:
      type: object
      properties:
        targetId:
          type: string
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        attachmentUrls:
          type: array
          items:
            type: string
      required: [targetId, body]
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        attachmentUrls:
          type: array
          items:
            type: string
...