# lib/api-zod/src/generated/types/createNoteRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberLab API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNoteRequest {
  targetId: string;
  title?: string;
  body: string;
  tags?: string[];
  attachmentUrls?: string[];
}