# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberLab API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from './findingSeverity';
import type { FindingStatus } from './findingStatus';

export interface Finding {
  id: string;
  userId: string;
  targetId: string;
  scanJobId?: string | null;
  title: string;
  cveId?: string | null;
  severity: FindingSeverity;
  status: FindingStatus;
  cvssScore?: number | null;
  owaspCategory?: string | null;
  description?: string | null;
  remediation?: string | null;
  references: string[];
  notes?: string | null;
  resolvedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}