# lib/api-zod/src/generated/types/updateFindingRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberLab API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateFindingRequestSeverity } from './updateFindingRequestSeverity';
import type { UpdateFindingRequestStatus } from './updateFindingRequestStatus';

export interface UpdateFindingRequest {
  status?: UpdateFindingRequestStatus;
  notes?: string;
  remediation?: string;
  severity?: UpdateFindingRequestSeverity;
  note?: string;
}